# Optimize json encoding