--- a/config/initializers/json.rb
+++ b/config/initializers/json.rb
@@ -1,7 +1,15 @@
 require 'active_support/json/encoding'
 
 module CCInitializers
-  def self.json(_cc_config)
+  def self.json(cc_config)
+    if cc_config[:use_optimized_json_encoder]
+      MultiJson.use(:oj)
+      Oj::Rails.optimize # Use optimized encoders instead of as_json() methods for available classes.
+      Oj.default_options = {
+        bigdecimal_load: :bigdecimal,
+      }
+    end
+
     ActiveSupport.json_encoder = Class.new do
       attr_reader :options
 
@@ -10,10 +18,16 @@ module CCInitializers
       end
 
       def encode(value)
-        if Rails.env.test?
-          MultiJson.dump(value.as_json(options.dup))
+        if MultiJson.default_adapter == :oj && value.is_a?(VCAP::CloudController::Presenters::V3::BasePresenter)
+          v = value.to_hash
         else
-          MultiJson.dump(value.as_json(options.dup), options.merge(pretty: true))
+          v = value.as_json(options.dup)
+        end
+
+        if Rails.env.test?
+          MultiJson.dump(v)
+        else
+          MultiJson.dump(v, options.merge(pretty: true))
         end
       end
     end