--- a/lib/cloud_controller/config_schemas/base/api_schema.rb
+++ b/lib/cloud_controller/config_schemas/base/api_schema.rb
@@ -354,6 +354,8 @@ module VCAP::CloudController
               write_key: String,
               dataset: String,
             },
+
+            optional(:use_optimized_json_encoder) => bool,
           }
         end
         # rubocop:enable Metrics/BlockLength