--- a/spec/acceptance/broker_api_compatibility/broker_api_versions_spec.rb
+++ b/spec/acceptance/broker_api_compatibility/broker_api_versions_spec.rb
@@ -17,7 +17,7 @@ RSpec.describe 'Broker API Versions' do
       'broker_api_v2.11_spec.rb' => '99e61dc50ceb635b09b3bd16901a4fa6',
       'broker_api_v2.12_spec.rb' => '4023dffdcaae014556dcdba9f7d206bb',
       'broker_api_v2.13_spec.rb' => 'ceecc106d1a203002e277a37812b6df3',
-      'broker_api_v2.14_spec.rb' => 'a1e7485793ba1916ea2f4080943530a5',
+      'broker_api_v2.14_spec.rb' => 'a9fbd1d91bcc9c33c054c7854be1ab5a',
       'broker_api_v2.15_spec.rb' => 'c575fd37bc6dc8df4f773719ccef3288',
     }
   end