While looking at the performance of the /v3/security_groups endpoints (PR #2475), I also checked how much the database queries were contributing to the overall response times. As this was not much after optimizing the queries, I started to look into the topic of JSON serialization. I found several comments and benchmarks claiming that [Oj](https://github.com/ohler55/oj) is the fastest JSON encoder/decoder; so I started with a small PoC.

I've managed to completely switch to Oj and run all unit tests successfully. For comparison of JSON outputs before and after the change, I use the [jd](https://github.com/josephburnett/jd) command-line tool.

I decided to introduce an optional config property (`use_optimized_json_encoder`). I've enabled it on a foundation and successfully run CATS.

Here are some results of locally executed performance tests (`X-Runtime` header):
| |GET /v3/security_groups?per_page=500 as admin *|GET /v3/security_groups/:guid as regular user|
|--|--|--|
|baseline|19.802s|0.195s|
|PR #2255|5.520s|0.158s|
|PR #2475|3.492s|0.029s|
|this PR|2.506s|0.025s|

\* each security group is assigned to 500 spaces

* [x] I have reviewed the [contributing guide](https://github.com/cloudfoundry/cloud_controller_ng/blob/main/CONTRIBUTING.md)

* [x] I have viewed, signed, and submitted the Contributor License Agreement

* [x] I have made this pull request to the `main` branch

* [x] I have run all the unit tests using `bundle exec rake`

* [x] I have run [CF Acceptance Tests](https://github.com/cloudfoundry/cloud_controller_ng#cf-acceptance-tests-cats) *

\* We cannot run all CATS on our foundations as we do not enable all CF features, but all the executed tests were successful:
```
Ran 136 of 216 Specs
SUCCESS! -- 136 Passed | 0 Failed | 3 Pending | 77 Skipped
```