@JenGoldstrich 2021-09-15
Hey @philippthun, 

Thanks for the PR!

Oj seems actively maintained so that sounds good, one thing I did notice is that this PR introduces another gem `rspec-json_expectations` which seems to not be actively maintained based on the last commit being from two years ago and [this comment from May 2021 attempting to reach out to maintainers](https://github.com/waterlink/rspec-json_expectations/pull/36#issuecomment-847229711).  I think we should continue to use `.includes json-sub-string` or switch to a more actively maintained JSON library.