@philippthun 2021-09-22
Hi @JenGoldstrich! Thanks for your review. The `rspec-json_expectations` gem is indeed not really needed. I've removed it from the PR and adapted the tests.