@tjvman 2021-09-29
@philippthun Are you planning to make a PR to capi-release to add the new config property?