@philippthun 2021-09-30
@tjvman I'm not sure about this. Ideally using the optimized json encoder should become the default. It might make sense to add the config option also to CAPI and then enable it on some foundations to see if everything is really compatible. And in the end this option could then be removed again...