@tjvman 2021-10-07
@philippthun Yeah, that makes sense to me for this kind of performance optimization thing - make it an opt-in at first, then make it the default. How long (time or number of releases) do you think we should leave it as optional

I don't see a way to enable the `use_optimized_json_encoder` besides editing the config via ssh after deploying, but I might be missing something. Assuming I'm not, then the option should be exposed in capi-release - otherwise it doesn't make sense to have it be a configurable thing (don't want to encourage users to live-edit the CC config)