@Gerg 2021-11-11
I think this is a great change and should provide API-wide performance improvements!